[chore] findPlansByUserId repository 위치 변경
