+++ src/main/java/project/tripplan/domain/plan/repository/PlanRepositoryCustomImpl.java
@@ -4,9 +4,13 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
 import com.querydsl.core.BooleanBuilder;
+import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -21,6 +25,7 @@ import project.tripplan.domain.plan.entity.Plan;
 import project.tripplan.domain.plan.entity.QPlan;
 import project.tripplan.domain.plan.entity.QPlanPlaceCategory;
 import project.tripplan.domain.plan.entity.QPlanTransportationCategory;
+import project.tripplan.domain.user.dto.UserPlanRes;
 import project.tripplan.domain.user.entity.QUser;
 
 @Slf4j
@@ -188,4 +193,39 @@ public class PlanRepositoryCustomImpl implements PlanRepositoryCustom {
 	public List<Plan> findHotPlacePlans(String placeName, int limit) {
 		return List.of();
 	}
+
+	@Override
+	public Page<UserPlanRes> findPlansByUserId(Long userId, Pageable pageable) {
+		List<UserPlanRes> content = qf
+			.select(Projections.constructor(
+				UserPlanRes.class,
+				plan.id,
+				plan.title,
+				plan.createdAt,
+				plan.imageUrl.as("thumbnail"),
+				Expressions.stringTemplate(
+					"group_concat(DISTINCT {0})",
+					planPlaceCategory.placeCategory.name
+				)
+				,
+				plan.status.stringValue()
+			))
+			.from(plan)
+			.leftJoin(plan.planPlaceCategories, planPlaceCategory)
+			.where(plan.user.id.eq(userId))
+			.orderBy(plan.createdAt.desc())
+			.groupBy(plan.id)
+			.offset(pageable.getOffset())
+			.limit(pageable.getPageSize())
+			.fetch();
+
+		long total = qf
+			.select(plan.count())
+			.from(plan)
+			.where(plan.user.id.eq(userId))
+			.fetchOne();
+
+		return new PageImpl<>(content, pageable, total);
+	}
+
 }
